+++ src/pages/Ranking/Ranking.module.css
@@ -1,15 +1,3 @@
 .sectionTitle {
   padding: var(--space-5, 20px) var(--container-m-md, 16px);
 }
-
-.container {
-  padding: 0 var(--container-m-sm, 12px);
-  margin-block-end: 60px;
-}
-
-.rankCardList {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-10, 40px);
-  margin-block-end: var(--space-10, 40px);
-}
